AbstractEntityService moved to core package
