+++ src/main/java/egl/client/service/model/core/ProfileService.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 
 import egl.client.model.core.profile.Profile;
 import egl.client.repository.core.profile.ProfileRepository;
-import egl.client.service.model.AbstractEntityService;
 import javafx.beans.property.Property;
 import javafx.beans.property.SimpleObjectProperty;
