+++ src/main/java/egl/client/service/model/global/GlobalTopicInfoService.java
@@ -3,7 +3,7 @@ package egl.client.service.model.global;
 import egl.client.model.core.topic.Topic;
 import egl.client.model.global.topic.GlobalTopicInfo;
 import egl.client.repository.global.topic.GlobalTopicInfoRepository;
-import egl.client.service.model.AbstractEntityService;
+import egl.client.service.model.core.AbstractEntityService;
 import org.springframework.stereotype.Service;
 
 @Service
