+++ src/main/java/egl/client/service/model/local/LocalTopicTasksService.java
@@ -3,7 +3,7 @@ package egl.client.service.model.local;
 import egl.client.model.core.topic.TopicTasks;
 import egl.client.model.core.topic.TopicType;
 import egl.client.repository.local.topic.LocalTopicTasksRepository;
-import egl.client.service.model.AbstractEntityService;
+import egl.client.service.model.core.AbstractEntityService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
